+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [3.0.2]
+
+### Fixed
+
+- Fix esbuild development change between server and client components.
+- Fix rollup development change between server and client components.
+- Fix CSS entries for webpack.
+
 ## [3.0.1]
 
 ### Fixed